requirements.txt
celery>=5.0
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
redis>=4.0
requests==2.32.3
stripe==11.5.0
openai>=1.65.5
pinecone==6.0.1
html2text==2025.4.15


# djangorestframework-authtoken>=1.2.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.122.0
ngrok
django-celery-beat==2.5.0 
django-prometheus>=2.2.0